give prometheus formatting its own file
