+++ ntp-proto/src/identifiers.rs
@@ -37,7 +37,7 @@ impl ReferenceId {
         *self == Self::KISS_RSTR
     }
 
-    pub fn to_bytes(self) -> [u8; 4] {
+    pub(crate) fn to_bytes(self) -> [u8; 4] {
         self.0.to_be_bytes()
     }
